byref() argument must be a ctypes instance

+++ libcudnn.py
@@ -534,8 +534,8 @@ def cudnnTransformTensor(handle, alpha, srcDesc, srcData, beta, destDesc, destDa
         Pointer to data of the tensor described by destDesc descriptor.
     """
 
-    status = _libcudnn.cudnnTransformTensor(handle, ctypes.byref(alpha), srcDesc,
-                                                srcData, ctypes.byref(beta),
+    status = _libcudnn.cudnnTransformTensor(handle, ctypes.byref(ctypes.c_float(alpha)), srcDesc,
+                                                srcData, ctypes.byref(ctypes.c_float(beta)),
                                                 destDesc, destData)
     cudnnCheckStatus(status)
 
@@ -575,8 +575,8 @@ def cudnnAddTensor(handle, mode, alpha, biasDesc, biasData, beta, srcDestDesc, s
         Pointer to data of the tensor described by srcDestDesc.
     """
 
-    status = _libcudnn.cudnnAddTensor(handle, mode, ctypes.byref(alpha), biasDesc,
-                                        biasData, ctypes.byref(beta),
+    status = _libcudnn.cudnnAddTensor(handle, mode, ctypes.byref(ctypes.c_float(alpha)), biasDesc,
+                                        biasData, ctypes.byref(ctypes.c_float(beta)),
                                         srcDestDesc, srcDestData)
     cudnnCheckStatus(status)
 
@@ -599,7 +599,7 @@ def cudnnSetTensor(handle, srcDesc, srcData, alpha):
         Value that all elements of the tensor will be set to.
     """
 
-    status = _libcudnn.cudnnSetTensor(handle, srcDesc, srcData, ctypes.byref(alpha))
+    status = _libcudnn.cudnnSetTensor(handle, srcDesc, srcData, ctypes.byref(ctypes.c_float(alpha)))
     cudnnCheckStatus(status)
 
 _libcudnn.cudnnScaleTensor.restype = int
@@ -623,7 +623,7 @@ def cudnnScaleTensor(handle, srcDesc, srcData, alpha):
         Value that all elements of the tensor will be scaled with.
     """
 
-    status = _libcudnn.cudnnScaleTensor(handle, srcDesc, srcData, ctypes.byref(alpha))
+    status = _libcudnn.cudnnScaleTensor(handle, srcDesc, srcData, ctypes.byref(ctypes.c_float(alpha)))
     cudnnCheckStatus(status)
 
 _libcudnn.cudnnCreateFilterDescriptor.restype = int
@@ -1047,10 +1047,10 @@ def cudnnConvolutionForward(handle, alpha, srcDesc, srcData, filterDesc, filterD
         Data pointer to GPU memory associated with the tensor descriptor destDesc.
     """
 
-    status = _libcudnn.cudnnConvolutionForward(handle, ctypes.byref(alpha), srcDesc, srcData,
+    status = _libcudnn.cudnnConvolutionForward(handle, ctypes.byref(ctypes.c_float(alpha)), srcDesc, srcData,
                                             filterDesc, filterData,
                                             convDesc, algo, workspace, workSpaceSizeInBytes,
-                                            ctypes.byref(beta), destDesc, destData)
+                                            ctypes.byref(ctypes.c_float(beta)), destDesc, destData)
     cudnnCheckStatus(status)
 
 _libcudnn.cudnnConvolutionBackwardBias.restype = int
@@ -1089,8 +1089,8 @@ def cudnnConvolutionBackwardBias(handle, alpha, srcDesc, srcData, beta, destDesc
         destDesc.
     """
 
-    status = _libcudnn.cudnnConvolutionBackwardBias(handle, ctypes.byref(alpha), srcDesc, srcData,
-                                                    ctypes.byref(beta), destDesc, destData)
+    status = _libcudnn.cudnnConvolutionBackwardBias(handle, ctypes.byref(ctypes.c_float(alpha)), srcDesc, srcData,
+                                                    ctypes.byref(ctypes.c_float(beta)), destDesc, destData)
     cudnnCheckStatus(status)
 
 _libcudnn.cudnnConvolutionBackwardFilter.restype = int
@@ -1135,9 +1135,9 @@ def cudnnConvolutionBackwardFilter(handle, alpha, srcDesc, srcData, diffDesc, di
         gradDesc that carries the result.
     """
 
-    status = _libcudnn.cudnnConvolutionBackwardFilter(handle, ctypes.byref(alpha), srcDesc,
+    status = _libcudnn.cudnnConvolutionBackwardFilter(handle, ctypes.byref(ctypes.c_float(alpha)), srcDesc,
                                                         srcData, diffDesc,
-                                                        diffData, convDesc, ctypes.byref(beta),
+                                                        diffData, convDesc, ctypes.byref(ctypes.c_float(beta)),
                                                         gradDesc, gradData)
     cudnnCheckStatus(status)
 
@@ -1184,9 +1184,9 @@ def cudnnConvolutionBackwardData(handle, alpha, filterDesc, filterData, diffDesc
         gradDesc that carries the result.
     """
 
-    status = _libcudnn.cudnnConvolutionBackwardData(handle, ctypes.byref(alpha), filterDesc,
+    status = _libcudnn.cudnnConvolutionBackwardData(handle, ctypes.byref(ctypes.c_float(alpha)), filterDesc,
                                                     filterData, diffDesc, diffData, convDesc,
-                                                    ctypes.byref(beta), gradDesc, gradData)
+                                                    ctypes.byref(ctypes.c_float(beta)), gradDesc, gradData)
     cudnnCheckStatus(status)
 
 _libcudnn.cudnnSoftmaxForward.restype = int
@@ -1223,8 +1223,8 @@ def cudnnSoftmaxForward(handle, algorithm, mode, alpha, srcDesc, srcData, beta, 
         destDesc.
     """
 
-    status = _libcudnn.cudnnSoftmaxForward(handle, algorithm, mode, ctypes.byref(alpha),
-                                        srcDesc, srcData, ctypes.byref(beta),
+    status = _libcudnn.cudnnSoftmaxForward(handle, algorithm, mode, ctypes.byref(ctypes.c_float(alpha)),
+                                        srcDesc, srcData, ctypes.byref(ctypes.c_float(beta)),
                                         destDesc, destData)
     cudnnCheckStatus(status)
 
@@ -1269,9 +1269,9 @@ def cudnnSoftmaxBackward(handle, algorithm, mode, alpha, srcDesc, srcData, srcDi
         destDiffDesc.
     """
 
-    status = _libcudnn.cudnnSoftmaxBackward(handle, algorithm, mode, ctypes.byref(alpha),
+    status = _libcudnn.cudnnSoftmaxBackward(handle, algorithm, mode, ctypes.byref(ctypes.c_float(alpha)),
                                             srcDesc, srcData,
-                                            srcDiffDesc, srcDiffData, ctypes.byref(beta),
+                                            srcDiffDesc, srcDiffData, ctypes.byref(ctypes.c_float(beta)),
                                             destDiffDesc, destDiffData)
     cudnnCheckStatus(status)
 
@@ -1429,8 +1429,8 @@ def cudnnPoolingForward(handle, poolingDesc, alpha, srcDesc, srcData, beta, dest
         destDesc.
     """
 
-    status = _libcudnn.cudnnPoolingForward(handle, poolingDesc, ctypes.byref(alpha),
-                                        srcDesc, srcData, ctypes.byref(beta),
+    status = _libcudnn.cudnnPoolingForward(handle, poolingDesc, ctypes.byref(ctypes.c_float(alpha)),
+                                        srcDesc, srcData, ctypes.byref(ctypes.c_float(beta)),
                                         destDesc, destData)
     cudnnCheckStatus(status)
 
@@ -1482,9 +1482,9 @@ def cudnnPoolingBackward(handle, poolingDesc, alpha, srcDesc, srcData, srcDiffDe
         destDiffDesc.
     """
 
-    status = _libcudnn.cudnnPoolingBackward(handle, poolingDesc, ctypes.byref(alpha),
+    status = _libcudnn.cudnnPoolingBackward(handle, poolingDesc, ctypes.byref(ctypes.c_float(alpha)),
                                             srcDesc, srcData, srcDiffDesc, srcDiffData,
-                                            destDesc, destData, ctypes.byref(beta),
+                                            destDesc, destData, ctypes.byref(ctypes.c_float(beta)),
                                             destDiffDesc, destDiffData)
     cudnnCheckStatus(status)
 
@@ -1528,8 +1528,8 @@ def cudnnActivationForward(handle, mode, alpha, srcDesc, srcData, beta, destDesc
         destDesc.
     """
 
-    status = _libcudnn.cudnnActivationForward(handle, mode, ctypes.byref(alpha), srcDesc, srcData,
-                                              ctypes.byref(beta), destDesc, destData)
+    status = _libcudnn.cudnnActivationForward(handle, mode, ctypes.byref(ctypes.c_float(alpha)), srcDesc, srcData,
+                                              ctypes.byref(ctypes.c_float(beta)), destDesc, destData)
     cudnnCheckStatus(status)
 
 _libcudnn.cudnnActivationBackward.restype = int
@@ -1584,8 +1584,8 @@ def cudnnActivationBackward(handle, mode, alpha, srcDesc, srcData, srcDiffDesc, 
         destDiffDesc.
     """
 
-    status = _libcudnn.cudnnActivationBackward(handle, mode, ctypes.byref(alpha), srcDesc, srcData,
+    status = _libcudnn.cudnnActivationBackward(handle, mode, ctypes.byref(ctypes.c_float(alpha)), srcDesc, srcData,
                                                srcDiffDesc, srcDiffData,
-                                               destDesc, destData, ctypes.byref(beta),
+                                               destDesc, destData, ctypes.byref(ctypes.c_float(beta)),
                                                destDiffDesc, destDiffData)
     cudnnCheckStatus(status)
